Allow usage of ResizeObserver in ResizeSensor
