+++ pubspec.yaml
@@ -22,12 +22,14 @@ dependencies:
   w_flux: ^2.7.1
   platform_detect: ^1.3.2
   quiver: ">=0.21.4 <0.26.0"
+  resize_observer_polyfill_wrapper: ^1.0.0
 dev_dependencies:
   build_runner: ^0.5.0
   built_value_generator: ^4.2.0
   coverage: ^0.7.2
   dart_dev: ^1.7.6
   dependency_validator: ^1.0.0
+  dart_to_js_script_rewriter: ^1.0.1
   mockito: ^0.11.0
   over_react_test: ^1.2.0
   source_gen: ^0.7.2+1
@@ -40,5 +42,6 @@ transformers:
       ]
   - test/pub_serve:
       $include: test/**_test{.*,}.dart
+  - dart_to_js_script_rewriter
   # Reminder: dart2js should come after any other transformers that touch Dart code
   - $dart2js
