+++ test/over_react_component_test.html
@@ -21,6 +21,7 @@ limitations under the License.
     <!-- javascript -->
     <script src="packages/react/react_with_addons.js"></script>
     <script src="packages/react/react_dom.js"></script>
+    <script src="packages/resize_observer_polyfill_wrapper/resize_observer_polyfill.js"></script>
 
     <link rel="x-dart-test"  href="over_react_component_test.dart">
     <script src="packages/test/dart.js"></script>
